Move process_comm() to a new file
Starting to split up the big utils file
