+++ gprofiler/profilers/ruby.py
@@ -16,7 +16,8 @@ from gprofiler.log import get_logger_adapter
 from gprofiler.merge import parse_one_collapsed_file
 from gprofiler.profilers.profiler_base import ProcessProfilerBase
 from gprofiler.profilers.registry import register_profiler
-from gprofiler.utils import pgrep_maps, process_comm, random_prefix, removed_path, resource_path, run_process
+from gprofiler.utils import pgrep_maps, random_prefix, removed_path, resource_path, run_process
+from gprofiler.utils.process import process_comm
 
 logger = get_logger_adapter(__name__)
